Split out resolvers into separate files
